[fix] #64 ApiGeatewayConvertService 비디오 의존성 분리
