+++ src/main/java/com/numble/team3/converter/domain/ConvertVideoUtils.java
@@ -1,0 +1,11 @@
+package com.numble.team3.converter.domain;
+
+import com.numble.team3.converter.application.request.CreateVideoDto;
+import com.numble.team3.converter.application.response.GetConvertVideoDto;
+import java.io.IOException;
+
+public interface ConvertVideoUtils {
+  void saveTempVideoForConvert(String filname, CreateVideoDto dto) throws IOException;
+
+  GetConvertVideoDto processConvertVideo(String filename) throws IOException;
+}
